kernelci.cli: add .base_api module for APICommand
Move APICommand to a new module kernelci.cli.base_api so that
dependencies such as cloudevents aren't imported in kernelci.cli.base.
This allows commands that don't use the API to be run without having
cloudevents (or other API-specific dependencies) installed and limits
the base module to only common base code for all the commands.

Update all the related tools using APICommand accordingly.
