+++ kernelci/cli/pubsub.py
@@ -7,7 +7,8 @@
 
 import sys
 
-from .base import APICommand, Args, sub_main
+from .base import Args, sub_main
+from .base_api import APICommand
 
 
 class cmd_subscribe(APICommand):  # pylint: disable=invalid-name
