+++ kernelci/cli/show.py
@@ -7,7 +7,8 @@
 
 from datetime import datetime
 
-from .base import APICommand, Args, sub_main
+from .base import Args, sub_main
+from .base_api import APICommand
 
 
 class cmd_results(APICommand):  # pylint: disable=invalid-name
